Add files via upload
This game you are fighting the forces of evil to try and save the world's programming languages from being extinguished only leaving the assembly language left. Essentially a rebranded version of hangman, but instead of drawing a little person, every time you get a wrong guess, you eliminate one of the programming languages from the world.